dpaa2: rc: extend ICID to 32-bits for VFIO support
While the ICID is only documented as a 16-bit value, 32-bits are
actually used (taking over the 'reserved' bits next to it).
The top 16 bits are then used to differentiate containers running
under a VM.

+++ sys/dev/dpaa2/dpaa2_mc.h
@@ -156,7 +156,7 @@ struct dpaa2_devinfo {
 
 	uint32_t		 id;
 	uint32_t		 portal_id;
-	uint16_t		 icid;
+	uint32_t		 icid;
 
 	enum dpaa2_dev_type	 dtype;
 	struct resource_list	 resources;
